move collectionPriceChart as shared component
